fix(ci): share chromeAvailable helper across chromedp and e2e build tags
Extract chromeAvailable() into helpers_test.go with build constraint
"chromedp || e2e" so both integration_test.go and e2e_test.go can use
it. Previously it was only in chrome_test.go (chromedp tag), causing
e2e builds to fail with "undefined: chromeAvailable".
